Bugfix: Do not account zIIPs in layer QC_LAYER_TYPE_ZOS_ZCX_SERVER
In a zIIPs-only zCX Server, STSI reports zIIPs as CPUs, and we initialize
all qc_num_cpu_* attributes in the layer with those values. However, once
STHYI data is evaluated, we'd update those values with the correct one:
zero. Hence we have to reset those attributes manually upon encountering
this case. To do so, we have to turn off the consistency check.

Signed-off-by: Stefan Raspl <raspl@linux.ibm.com>