qc_open: Fix memory leak on error
When calling qc_open() results in an error, implicit cleanup via qc_close()
is aborted, as the handle is not registered.
To fix, we simply register a handle as soon as we get it. There is still a
pathological case where we cannot register the handle - but there's nothing
we can do about that case. We leave the handle set, still.

Signed-off-by: Stefan Raspl <raspl@linux.ibm.com>