Add new attributes in support of secure boot in all final layers
Add two attributes that will indicate whether secure boot is available and
in use.
Note that we will only ever get the values for the final layer, as documented
in the README entry. Otherwise, we would need to have the values provided via
e.g. STHYI for further layers - which is currently not architectured.
Might lead to some confusion, when the topmost layer indicates support,
but the lower layers have no info available. Was thinking about post-
processing the data into lower layers, but that won't work at all: In case
secure boot is available but not set, we cannot know at which of the lower
layers things broke.

Signed-off-by: Stefan Raspl <raspl@linux.ibm.com>